+++ _vendor/python_rapidjson/rapidjson.cpp
@@ -230,10 +230,15 @@ static PyObject* encoder_new(PyTypeObject* type, PyObject* args, PyObject* kwarg
 static PyObject* validator_call(PyObject* self, PyObject* args, PyObject* kwargs);
 static void validator_dealloc(PyObject* self);
 static PyObject* validator_new(PyTypeObject* type, PyObject* args, PyObject* kwargs);
+static PyObject* validator_validate(PyObject* self, PyObject* args, PyObject* kwargs);
+static PyObject* validator_check_schema(PyObject* cls, PyObject* args, PyObject* kwargs);
 
 static PyObject* normalizer_call(PyObject* self, PyObject* args, PyObject* kwargs);
 static void normalizer_dealloc(PyObject* self);
 static PyObject* normalizer_new(PyTypeObject* type, PyObject* args, PyObject* kwargs);
+static PyObject* normalizer_validate(PyObject* self, PyObject* args, PyObject* kwargs);
+static PyObject* normalizer_check_schema(PyObject* cls, PyObject* args, PyObject* kwargs);
+static PyObject* normalizer_normalize(PyObject* self, PyObject* args, PyObject* kwargs);
 
 ///////////////////////////////////////////////////
 // Stream wrapper around Python file-like object //
@@ -4302,6 +4307,18 @@ PyDoc_STRVAR(validator_doc,
              " string or Python dictionary.");
 
 
+static PyMethodDef validator_methods[] = {
+    {"validate", (PyCFunction) validator_validate,
+     METH_VARARGS | METH_KEYWORDS,
+     "Validate a JSON document."
+    },
+    {"check_schema", (PyCFunction) validator_check_schema,
+     METH_VARARGS | METH_KEYWORDS | METH_CLASS,
+     "Validate a schema against the JSON metaschema."},
+    {NULL}  /* Sentinel */
+};
+
+
 static PyTypeObject Validator_Type = {
     PyVarObject_HEAD_INIT(NULL, 0)
     "rapidjson.Validator",          /* tp_name */
@@ -4330,7 +4347,7 @@ static PyTypeObject Validator_Type = {
     0,                              /* tp_weaklistoffset */
     0,                              /* tp_iter */
     0,                              /* tp_iternext */
-    0,                              /* tp_methods */
+    validator_methods,              /* tp_methods */
     0,                              /* tp_members */
     0,                              /* tp_getset */
     0,                              /* tp_base */
@@ -4421,7 +4438,6 @@ static PyObject* validator_call(PyObject* self, PyObject* args, PyObject* kwargs
     Py_RETURN_NONE;
 }
 
-
 static void validator_dealloc(PyObject* self)
 {
     ValidatorObject* s = (ValidatorObject*) self;
@@ -4464,7 +4480,7 @@ static PyObject* validator_new(PyTypeObject* type, PyObject* args, PyObject* kwa
         NULL
     };
 
-    if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O|$OOOOOOOp:Normalizer",
+    if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O|$OOOOOOOp:Validator",
                                      (char**) kwlist,
 				     &jsonObject,
                                      &objectHook,
@@ -4558,6 +4574,158 @@ static PyObject* validator_new(PyTypeObject* type, PyObject* args, PyObject* kwa
 }
 
 
+static PyObject* validator_validate(PyObject* self, PyObject* args, PyObject* kwargs)
+{ return validator_call(self, args, kwargs); }
+
+
+static PyObject* validator_check_schema(PyObject* cls, PyObject* args, PyObject* kwargs)
+{
+    PyObject* jsonObject;
+    PyObject* objectHook = NULL;
+    PyObject* numberModeObj = NULL;
+    unsigned numberMode = NM_NAN;
+    PyObject* datetimeModeObj = NULL;
+    unsigned datetimeMode = DM_NONE;
+    PyObject* uuidModeObj = NULL;
+    unsigned uuidMode = UM_NONE;
+    PyObject* bytesModeObj = NULL;
+    unsigned bytesMode = BM_UTF8;
+    PyObject* iterableModeObj = NULL;
+    unsigned iterableMode = IM_ANY_ITERABLE;
+    PyObject* mappingModeObj = NULL;
+    unsigned mappingMode = MM_ANY_MAPPING;
+    int allowNan = -1;
+    static char const* kwlist[] = {
+        "object_hook"
+        "number_mode",
+        "datetime_mode",
+        "uuid_mode",
+        "bytes_mode",
+        "iterable_mode",
+        "mapping_mode",
+
+        /* compatibility with stdlib json */
+        "allow_nan",
+
+        NULL
+    };
+
+    if (!PyArg_ParseTupleAndKeywords(args, kwargs,
+				     "O|$OOOOOOOp:Validator.check_schema",
+                                     (char**) kwlist,
+				     &jsonObject,
+                                     &objectHook,
+                                     &numberModeObj,
+                                     &datetimeModeObj,
+                                     &uuidModeObj,
+				     &bytesModeObj,
+				     &iterableModeObj,
+				     &mappingModeObj,
+                                     &allowNan))
+        return NULL;
+
+    const char* jsonStr;
+
+    if (PyBytes_Check(jsonObject)) {
+        jsonStr = PyBytes_AsString(jsonObject);
+        if (jsonStr == NULL)
+            return NULL;
+    } else if (PyUnicode_Check(jsonObject)) {
+        jsonStr = PyUnicode_AsUTF8(jsonObject);
+        if (jsonStr == NULL)
+            return NULL;
+    } else if (PyDict_Check(jsonObject)) {
+        jsonStr = NULL;
+    } else {
+        PyErr_SetString(PyExc_TypeError, "Expected string or UTF-8 encoded bytes");
+        return NULL;
+    }
+
+    if (objectHook && !PyCallable_Check(objectHook)) {
+        if (objectHook == Py_None) {
+            objectHook = NULL;
+        } else {
+            PyErr_SetString(PyExc_TypeError, "object_hook is not callable");
+            return NULL;
+        }
+    }
+
+    if (!accept_number_mode_arg(numberModeObj, allowNan, numberMode))
+        return NULL;
+
+    if (!accept_datetime_mode_arg(datetimeModeObj, datetimeMode))
+        return NULL;
+
+    if (!accept_uuid_mode_arg(uuidModeObj, uuidMode))
+        return NULL;
+
+    if (!accept_bytes_mode_arg(bytesModeObj, bytesMode))
+        return NULL;
+
+    if (!accept_iterable_mode_arg(iterableModeObj, iterableMode))
+        return NULL;
+
+    if (!accept_mapping_mode_arg(mappingModeObj, mappingMode))
+        return NULL;
+
+    Document d;
+    bool error;
+
+    if (jsonStr == NULL) {
+        error = (!PythonAccept(&d, jsonObject, numberMode, datetimeMode,
+			       uuidMode, bytesMode, iterableMode, mappingMode));
+	d.FinalizeFromStack();
+    } else {
+        Py_BEGIN_ALLOW_THREADS
+        error = d.Parse(jsonStr).HasParseError();
+        Py_END_ALLOW_THREADS
+    }
+    
+    if (error) {
+        PyErr_SetString(decode_error, "Invalid JSON");
+        return NULL;
+    }
+
+    Document d_meta;
+    error = d_meta.Parse("{\"type\": \"schema\"}").HasParseError();
+    if (error) {
+	PyErr_SetString(decode_error, "Invalid metaschema");
+	return NULL;
+    }
+
+    SchemaDocument metaschema(d_meta);
+    SchemaValidator validator(metaschema);
+    bool accept;
+    
+    Py_BEGIN_ALLOW_THREADS
+    accept = d.Accept(validator);
+    Py_END_ALLOW_THREADS
+
+    if (!accept) {
+        StringBuffer sptr;
+        StringBuffer dptr;
+
+        Py_BEGIN_ALLOW_THREADS
+        validator.GetInvalidSchemaPointer().StringifyUriFragment(sptr);
+        validator.GetInvalidDocumentPointer().StringifyUriFragment(dptr);
+        Py_END_ALLOW_THREADS
+
+        PyObject* error = Py_BuildValue("sss", validator.GetInvalidSchemaKeyword(),
+                                        sptr.GetString(), dptr.GetString());
+        PyErr_SetObject(validation_error, error);
+
+        Py_XDECREF(error);
+        sptr.Clear();
+        dptr.Clear();
+
+        return NULL;
+    }
+
+    Py_RETURN_NONE;
+    
+}
+
+
 ////////////////
 // Normalizer //
 ////////////////
@@ -4584,6 +4752,20 @@ PyDoc_STRVAR(normalizer_doc,
              "\n"
              "Create and return a new Normalizer instance from the given `json_schema`"
              " string or Python dictionary.");
+
+
+static PyMethodDef normalizer_methods[] = {
+    {"validate", (PyCFunction) normalizer_validate,
+     METH_VARARGS | METH_KEYWORDS,
+     "Validate a JSON document."},
+    {"normalize", (PyCFunction) normalizer_normalize,
+     METH_VARARGS | METH_KEYWORDS,
+     "Normalize a JSON document."},
+    {"check_schema", (PyCFunction) normalizer_check_schema,
+     METH_VARARGS | METH_KEYWORDS | METH_CLASS,
+     "Validate a schema against the JSON metaschema."},
+    {NULL}  /* Sentinel */
+};
 
 
 static PyTypeObject Normalizer_Type = {
@@ -4614,7 +4796,7 @@ static PyTypeObject Normalizer_Type = {
     0,                              /* tp_weaklistoffset */
     0,                              /* tp_iter */
     0,                              /* tp_iternext */
-    0,                              /* tp_methods */
+    normalizer_methods,             /* tp_methods */
     0,                              /* tp_members */
     0,                              /* tp_getset */
     0,                              /* tp_base */
@@ -4853,6 +5035,91 @@ static PyObject* normalizer_new(PyTypeObject* type, PyObject* args, PyObject* kw
 
     return (PyObject*) v;
 }
+
+
+static PyObject* normalizer_normalize(PyObject* self, PyObject* args, PyObject* kwargs)
+{ return normalizer_call(self, args, kwargs); }
+
+
+static PyObject* normalizer_validate(PyObject* self, PyObject* args, PyObject* kwargs)
+{
+    PyObject* jsonObject;
+
+    if (!PyArg_ParseTuple(args, "O", &jsonObject))
+        return NULL;
+
+    const char* jsonStr;
+    Py_ssize_t jsonStrLen = 0;
+
+    if (PyBytes_Check(jsonObject)) {
+        jsonStr = PyBytes_AsString(jsonObject);
+        if (jsonStr == NULL)
+            return NULL;
+	jsonStrLen = PyBytes_Size(jsonObject);
+    } else if (PyUnicode_Check(jsonObject)) {
+        jsonStr = PyUnicode_AsUTF8AndSize(jsonObject, &jsonStrLen);
+        if (jsonStr == NULL)
+            return NULL;
+    } else {
+        jsonStr = NULL;
+    }
+
+    if ((jsonStr != NULL) && (!isJSONDocument(jsonStr, jsonStrLen)))
+	jsonStr = NULL;
+
+    Document d;
+    bool error;
+    ValidatorObject* v = (ValidatorObject*) self;
+
+    if (jsonStr == NULL) {
+        error = (!PythonAccept(&d, jsonObject, v->numberMode, v->datetimeMode,
+			       v->uuidMode, v->bytesMode, v->iterableMode,
+			       v->mappingMode));
+	d.FinalizeFromStack();
+    } else {
+        Py_BEGIN_ALLOW_THREADS
+        error = d.Parse(jsonStr).HasParseError();
+        Py_END_ALLOW_THREADS
+    }
+
+    if (error) {
+        PyErr_SetString(decode_error, "Invalid JSON");
+        return NULL;
+    }
+
+    SchemaValidator validator(*((NormalizerObject*) self)->schema);
+    bool accept;
+
+    Py_BEGIN_ALLOW_THREADS
+    accept = d.Accept(validator);
+    Py_END_ALLOW_THREADS
+
+    if (!accept) {
+        StringBuffer sptr;
+        StringBuffer dptr;
+
+        Py_BEGIN_ALLOW_THREADS
+        validator.GetInvalidSchemaPointer().StringifyUriFragment(sptr);
+        validator.GetInvalidDocumentPointer().StringifyUriFragment(dptr);
+        Py_END_ALLOW_THREADS
+
+        PyObject* error = Py_BuildValue("sss", validator.GetInvalidSchemaKeyword(),
+                                        sptr.GetString(), dptr.GetString());
+        PyErr_SetObject(validation_error, error);
+
+        Py_XDECREF(error);
+        sptr.Clear();
+        dptr.Clear();
+
+        return NULL;
+    }
+
+    Py_RETURN_NONE;
+}
+
+
+static PyObject* normalizer_check_schema(PyObject* cls, PyObject* args, PyObject* kwargs)
+{ return validator_check_schema(cls, args, kwargs); }
 
 
 ////////////